+++ tests/ats/setup-local.sh
@@ -16,5 +16,7 @@ SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
   helm uninstall -n default konfigure-operator > /dev/null || true
 
   echo "Installing new konfigure-operator helm release in default namespace"
-  helm install -n default -n default --set "image.tag=black" konfigure-operator ./helm/konfigure-operator
+  helm install -n default -n default --set "image.tag=black,image.pullPolicy=IfNotPresent" konfigure-operator ./helm/konfigure-operator
+
+  kubectl rollout restart -n default deployment konfigure-operator
 )